kexec: different type comparison in crash kernel
min() is strict regarding the types of its operands.

+++ arch/powerpc/kernel/crash_dump.c
@@ -128,7 +128,7 @@ ssize_t copy_oldmem_page(unsigned long pfn, char *buf,
 	if (!csize)
 		return 0;
 
-	csize = min(csize, PAGE_SIZE);
+	csize = min(csize, (size_t)PAGE_SIZE);
 
 	if (pfn < max_pfn) {
 		vaddr = __va(pfn << PAGE_SHIFT);
